backend/requirements.txt
# Web Framework and Server
fastapi
uvicorn[standard]

# Data validation (used by FastAPI)
pydantic

# For multipart/form-data (file uploads)
python-multipart

# Google Cloud & Firebase
firebase-admin
google-api-python-client
google-auth-oauthlib

# For loading environment variables from a .env file
python-dotenv